Methods
Using PrivateJet class, variables were declared. Then in the main method a PrivateJet object was created. Two methods were created in the PrivateJet class, one to output the description of the jet. The other method outputs how many of these jets are for sale. The isAvailable method is called inside an if statement and is only executed if the boolean "forSale" = true, the state which is randomized each time the program is ran.